Split flatten operator sources.
